requirements.txt
Flask==2.3.3
opencv-python==4.8.0.76
deepface==0.0.79
Pillow==10.0.1
numpy==1.24.3
tf-keras==2.13.0
tensorflow==2.13.0